Product Details page Done.

+++ app/src/main/java/com/aakash/ecommerce/repository/CategoryRepository.kt
@@ -44,7 +44,7 @@ class CategoryRepository @Inject constructor(
             .addOnSuccessListener { documents ->
                 val products = documents.map { document ->
                     Products(
-                        id = document.id.toInt(),
+                        id = document.id,
                         productName = document.getString("title") ?: "",
                         productPrice = document.getDouble("price") ?: 0.0,
                         productImageUrl = document.getString("imageUrl") ?: ""
